Optimize entity retrieval with classifier filters
Skips package filter when package not defined to save on possibly expensive regex matching.
This changes behaviour in that if a package is not defined, we will
retrieve all entities as opposed to no entities.

Improve efficiency of classifier filtering by executing filter on db
index instead of in memory filter.

Clarify entities by package api description.

Optimizes imports.